+++ apps/web/app/admin/members/page.tsx
@@ -1,7 +1,7 @@
+import AsyncBoundary from "@/components/common/Boundary";
+import ErrorFallback from "@/components/common/Fallback";
 import MembersContent from "./_components/MembersContent";
 import Skeleton from "./_components/skeleton";
-import AsyncBoundary from "./_components/AsyncBoundary";
-import ErrorFallback from "./_components/ErrorFallback";
 
 export default function MembersPage(): JSX.Element {
   return (
